+++ src/importsource.cpp
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include "libcellml/import.h"
+#include "libcellml/importsource.h"
 
 namespace libcellml {
